Use WorkManager handle delete account requests.  Split apart the logic for signing out of the device and out of the server.
