+++ EmbeddedSocialClient/sdk/src/main/java/com/microsoft/embeddedsocial/service/worker/DeleteAccountWorker.java
@@ -3,34 +3,41 @@
  * Licensed under the MIT License. See LICENSE in the project root for license information.
  */
 
-package com.microsoft.embeddedsocial.service.handler;
+package com.microsoft.embeddedsocial.service.worker;
 
 import com.microsoft.embeddedsocial.account.UserAccount;
-import com.microsoft.embeddedsocial.actions.Action;
 import com.microsoft.embeddedsocial.base.GlobalObjectRegistry;
 import com.microsoft.embeddedsocial.base.utils.debug.DebugLog;
 import com.microsoft.embeddedsocial.server.EmbeddedSocialServiceProvider;
 import com.microsoft.embeddedsocial.server.IAccountService;
 import com.microsoft.embeddedsocial.server.exception.NetworkRequestException;
 import com.microsoft.embeddedsocial.server.model.account.DeleteUserRequest;
-import com.microsoft.embeddedsocial.service.ServiceAction;
 
-import android.content.Intent;
+import android.content.Context;
+
+import androidx.work.Worker;
+import androidx.work.WorkerParameters;
 
 /**
- * Sends a delete account request to the server.
+ * Sends a delete account request to the server
  */
-public class DeleteAccountHandler extends ActionHandler {
+public class DeleteAccountWorker extends Worker {
+
+    public DeleteAccountWorker(Context context, WorkerParameters workerParameters) {
+        super(context, workerParameters);
+    }
+
     @Override
-    protected void handleAction(Action action, ServiceAction serviceAction, Intent intent) {
+    public Result doWork() {
         IAccountService server = GlobalObjectRegistry.getObject(EmbeddedSocialServiceProvider.class).getAccountService();
         try {
             server.deleteUser(new DeleteUserRequest());
-            UserAccount.getInstance().signOut();
-            action.complete();
+            UserAccount.getInstance().signOutOfDevice();
         } catch (NetworkRequestException e) {
             DebugLog.logException(e);
-            action.fail();
+            return Result.failure();
         }
+
+        return Result.success();
     }
 }
